# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Inventory management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List items with optional filters
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: locationId
          in: query
          required: false
          schema:
            type: integer
        - name: includeSubcategories
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      operationId: createItem
      tags: [items]
      summary: Create a new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get an item by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateItem
      tags: [items]
      summary: Update an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemBody"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Item:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        quantity:
          type: integer
        photoUrl:
          type: ["string", "null"]
        unitId:
          type: ["integer", "null"]
        unitSymbol:
          type: ["string", "null"]
        unitName:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        userId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - description
        - quantity
        - unit
        - location
        - sku
        - barcode
        - tags
        - notes
        - categoryId
        - categoryName
        - userId
        - createdAt
        - updatedAt
    CreateItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        quantity:
          type: integer
        photoUrl:
          type: ["string", "null"]
        unitId:
          type: ["integer", "null"]
        unit:
          type: ["string", "null"]
        locationId:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
      required:
        - name
        - quantity
    UpdateItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        quantity:
          type: integer
        photoUrl:
          type: ["string", "null"]
        unitId:
          type: ["integer", "null"]
        unit:
          type: ["string", "null"]
        locationId:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
...