# lib/api-zod/src/generated/types/updateItemBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Inventory management API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateItemBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  quantity?: number;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  unitId?: number | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  locationId?: number | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  barcode?: string | null;
  /** @nullable */
  tags?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  categoryId?: number | null;
}